Prise en compte des erreurs 404 de notre index SIRENE (#2255)
* do not make call to INSEE API when siret is not found
* do not import mutation from other test
