+++ back/src/forms/resolvers/mutations/__tests__/mutations.ts
@@ -1,0 +1,29 @@
+export const MARK_AS_SEALED = `
+  mutation MarkAsSealed($id: ID!) {
+    markAsSealed(id: $id) {
+      id
+      status
+    }
+  }
+`;
+
+export const SIGN_EMISSION_FORM = `
+  mutation SignEmissionForm($id: ID!, $input: SignEmissionFormInput!, $securityCode: Int) {
+    signEmissionForm(id: $id, input: $input, securityCode: $securityCode) {
+      id
+      status
+      signedByTransporter
+      sentAt
+      sentBy
+      emittedAt
+      emittedBy
+      emittedByEcoOrganisme
+      temporaryStorageDetail {
+        signedAt
+        signedBy
+        emittedAt
+        emittedBy
+      }
+    }
+  }
+`;
